Split SDLSurface to public class Add SDLActivity method createSDLSurface
